requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
bcrypt>=4.2.0
itsdangerous>=2.2.0
python-multipart>=0.0.12
pdfplumber>=0.11.0
python-docx>=1.1.0
anthropic>=0.40.0
jinja2>=3.1.4
python-dotenv>=1.0.1
